fix: share Facebook login scopes
